requirements.txt
aerich
fastapi
uvicorn
aiohttp
aiosqlite
cachetools
coloredlogs
redis
toml
tortoise-orm
requests
gitpython
nest-asyncio
dill
nltk
